Improves performance by refactoring the way we use the Mirror component
1. We now apply the `Mirror` component to individual sections instead of
   around the entire page
2. Parallax effect of the mirror is now removed
3. Performance is better. Perhaps not by much, because the blur is still
   present, but this should already help.

In the previous version we were rendering the `Mirror` around the entire
page, and then cherry-picking elements to hide, for performance or
visual reasons.

Most of the hidden mirror components would need `visibility: hidden`
 instead of `display: none`, to preserve height. So they end up being
 rendered anyway, which still hurts performance, especially for images

Now, a new approach is used, where the global mirror is removed, and
instead added to specific places where we actually want it to show up.

These places are mostly title and text components. Since mirrors are
applied per section, we can now freely hide images with `display: none`,
as long as they're at the end of their container, which is the case for
all sections so far.

One drawback of this approach is that I had to remove the parallax
effect when moving the mouse. This drawback is another performance improvement in itself, so I still consider this a victory.